+++ era-compiler-solidity/src/project/mod.rs
@@ -17,7 +17,7 @@ use rayon::iter::ParallelIterator;
 
 use crate::build_eravm::Build as EraVMBuild;
 use crate::build_evm::Build as EVMBuild;
-use crate::missing_libraries::MissingLibraries;
+use crate::libraries::Libraries;
 use crate::process::input_eravm::dependency_data::DependencyData as EraVMProcessInputDependencyData;
 use crate::process::input_eravm::Input as EraVMProcessInput;
 use crate::process::input_evm::dependency_data::DependencyData as EVMProcessInputDependencyData;
@@ -441,8 +441,12 @@ impl Project {
 
         let mut hashes = HashMap::with_capacity(results.len());
         for (path, result) in results.iter() {
-            if let Ok(ref contract) = result {
-                hashes.insert(path.to_owned(), contract.build.bytecode_hash.to_owned());
+            if let Some(bytecode_hash) = result
+                .as_ref()
+                .ok()
+                .and_then(|contract| contract.build.bytecode_hash)
+            {
+                hashes.insert(path.to_owned(), bytecode_hash.to_owned());
             }
         }
 
@@ -457,16 +461,12 @@ impl Project {
                             .unwrap_or_else(|| {
                                 panic!("dependency `{dependency}` full path not found")
                             });
-                        let hash = hashes
-                            .get(dependency_path.as_str())
-                            .cloned()
-                            .unwrap_or_else(|| {
-                                panic!("dependency `{dependency_path}` not found in the project")
-                            });
-                        contract
-                            .build
-                            .factory_dependencies
-                            .insert(hex::encode(hash), dependency_path);
+                        if let Some(hash) = hashes.get(dependency_path.as_str()) {
+                            contract
+                                .build
+                                .factory_dependencies
+                                .insert(hex::encode(hash), dependency_path);
+                        }
                     }
                 }
                 (path, result)
@@ -510,7 +510,7 @@ impl Project {
     ///
     /// Get the list of missing deployable libraries.
     ///
-    pub fn get_missing_libraries(&self) -> MissingLibraries {
+    pub fn get_missing_libraries(&self) -> Libraries {
         let deployed_libraries = self
             .libraries
             .iter()
@@ -531,6 +531,6 @@ impl Project {
                 .collect::<HashSet<String>>();
             missing_deployable_libraries.insert(contract_path.to_owned(), missing_libraries);
         }
-        MissingLibraries::new(missing_deployable_libraries)
+        Libraries::new(missing_deployable_libraries)
     }
 }