+++ era-compiler-solidity/src/yul/parser/statement/assignment.rs
@@ -5,10 +5,10 @@
 use era_compiler_llvm_context::IContext;
 use inkwell::types::BasicType;
 
-use crate::create_wrapper;
+use crate::declare_wrapper;
 use crate::yul::parser::wrapper::Wrap;
 
-create_wrapper!(
+declare_wrapper!(
     era_yul::yul::parser::statement::assignment::Assignment,
     Assignment
 );