+++ era-compiler-solidity/src/yul/parser/statement/code.rs
@@ -1,12 +1,12 @@
 //!
-//! The YUL code.
+//! The Yul code.
 //!
 
-use crate::create_wrapper;
+use crate::declare_wrapper;
 use crate::yul::parser::dialect::era::EraDialect;
 use crate::yul::parser::wrapper::Wrap;
 
-create_wrapper!(
+declare_wrapper!(
     era_yul::yul::parser::statement::code::Code<EraDialect>,
     Code
 );