+++ era-compiler-solidity/src/yul/parser/statement/for_loop.rs
@@ -4,13 +4,13 @@
 
 use era_compiler_llvm_context::IContext;
 
-use crate::create_wrapper;
+use crate::declare_wrapper;
 use crate::yul::parser::dialect::era::EraDialect;
 use crate::yul::parser::wrapper::Wrap;
 
 use super::expression::Expression;
 
-create_wrapper!(
+declare_wrapper!(
     era_yul::yul::parser::statement::for_loop::ForLoop<EraDialect>,
     ForLoop
 );