+++ era-compiler-solidity/src/yul/parser/statement/mod.rs
@@ -2,7 +2,7 @@
 //! The block statement.
 //!
 
-use crate::create_wrapper;
+use crate::declare_wrapper;
 
 use super::dialect::era::EraDialect;
 
@@ -17,7 +17,7 @@ pub mod object;
 pub mod switch;
 pub mod variable_declaration;
 
-create_wrapper!(
+declare_wrapper!(
     era_yul::yul::parser::statement::Statement<EraDialect>,
     Statement
 );